🎁 Add campus to breadcrumbs
This commit will add relevant SVG files for the breadcrumbs and a few
helpers as well as some SolrDocument methods that we'll need.  The main
thing is adding the Ngao::Arclight::BreadcrumbsHierarchyComponent
because we want to override the breadcrumbs_hierarchy_component view.
Overriding view component views are a little different that regular view
files where if you match the path you'll be set.  This required us to
create the component object in this application and then it would be
able to pick up the overridden view component erb.

Ref:
- https://github.com/ViewComponent/view_component/issues/411