🎁 Add campuses to index page
This commit will add the campuses to the index page and do some minimum
styling.  Also the scroll to top button was added as well all its
accompanying files.  The repositories.yml was updated to NGAO's current
one so we can see more than one campus.  We also introduced a decorator
pattern to override the repositories controller to add the campus logic.