💄 Make breadcrumb styling padding left more even
Prior to this commit, the padding left was inconsistent and when we got
deep into the nested breadcrumbs, the padding would not account for the
situtation.  This commit will make the styling a little more dynamic in
that it will adjust the padding based on the number of nested
`.breadcrumb-item-3` elements.  The draw back is that you still would
need to adjust the number of nested elements if things get super deep.
Right now I set it to 8, anything beyond that would look wonky and would
require the upper number to be bumped up.