⚙️ Configure `ead2_config` and `CatalogController`
This commit brings over a lot from NGAO but I hesitate to copy it
directly since a lot of the changes are either not relevant or are
already the way Arclight does things.  I did the logic for the
`context.clipboard[:repository]` in the `ead2_config` file to look for
the campus if the REPOSITORY_ID wasn't provided which might be useful in
this time of transition.  This could be reverted back after we hand this
project back to IU.

A lot of the issues that came from the reviewing of the EAD
implementation should be fixed in this commit.

Ref:
- https://github.com/scientist-softserv/archives_online/issues/45
- https://github.com/scientist-softserv/archives_online/issues/46
- https://github.com/scientist-softserv/archives_online/issues/47
- https://github.com/scientist-softserv/archives_online/issues/48