💄 Style Arclight search breadcrumb
This will add some styling to the search breadcrumb that you see on the
homepage and search results page.  The styling mimics the Rivet
breadcrumb style.  A hacky approach to keeping the screen reader working
is moving the text off screen.  Hopefully this will pass accessibility
testing.  If not then a different approach will be needed.