💄 Style Repositories and Collections link
This commit will add the Rivet style to the Repositories and Collections
link in the header.  There is a subtle color shift when the link is
active but it points out perhaps a bug where if you choose Collection,
the active class doesn't get applied to the Collection <li> element.
This is an upstream issue that should be fixed.