💄 Add styling to search widgets
This commit adds the Rivet styles to the search widgets on the search
results page.  I decided against using the Rivet dropdown button because
it looked a little too out of place.  Instead it just uses the Rivet
secondary button with the recommended chevron-down icon.  The dropdown
menu, however, does use the Rivet style.