♿️ Override Range Limit gem for accessibility
This commit will override the Blacklight Range Limit v8.5.0 to swap the
submit button in the div.visually-hidden with the submit buton that is
not in the div.visually-hidden.  Also adding tabindex="-1" to the hidden
one. This will pass the Siteimprove Accessibility Checker.

The entire RangeFormComponent class had to be brought over because in
view components, we can't just bring over the view we want to override.
I am adding the class to the rubocop ignore because I don't want to
alter the contents since it's only here to support its view file to be
loaded.

Ref:
- https://github.com/scientist-softserv/archives_online/issues/68