⚙️ Revert the Arclight route override
This commit will revert a previous override of the Arclight route that
changes the default params from `level` to `level_ssim`.  That solved
the issue for the `Collections` link pointing to the correct faceted
search, it broke all other facet searches.  If I changed the route
directly in Arclight to add the `level_ssim` param, it worked fine.  If
I brought it in locally then the other facets break.  I'm not sure why
it was doing that but with regards to time, I going to override the link
of the `Collections` to the correct faceted search.