🎁 Exposing Collection ID Number
This commit moves the Collection ID number from the `More Info` dropdown
to the Collection information box on the right side of the page.

This change requires an override of the `collection_context_component`
and `collection_info_component` components, as well as the
`sidebar_component.rb` which renders the both of the components.

Ref:
- https://github.com/notch8/archives_online/issues/101