🚧 Modify print.scss and ead_processor.rb
This commit will alter the original print.scss and ead_processor.rb from
Archives Online main.  The screen styles weren't getting applied so I'm
moving it into the base ead-print style.  Also, adding some logic in the
EadProcessor to ensure that we are getting the same html file name as
the id of the document.