🎁 Adds container info to collection context
This commit adds the physical container information to the collection
context in the sidebar. It is justified to the right on the last line of
the title. The sidebar has been expanded to 5 grid columns and and the
main container has been reduced to 7 grid columns to accommodate the
container information.

Ref:
- https://github.com/notch8/archives_online/issues/103