🎁 Adds masthead text link and updates header link
This commit changes the link on the Indiana University header to be an
external link to the Indiana University homepage. It also adds a link to
the masthead text that will navigate back to the homepage of Archives
Online. Changing the masthead required overriding the header component
and bringing in the masthead components.

Ref:
- https://github.com/notch8/archives_online/issues/184