🐛 Custom labels for <odd> EAD tag item level
This commit adds a decorator for the metadata field layout component to
render custom labels. The initializer on MetadataFieldLayoutComponent
is not parsing some of the keys which was allowing the Solr label to
get passed into the key. This fix is checking for the presence of a
label, and if it not present, manually splitting the key to get the
correct key. There may be a future fix within the EAD2 configurations.

Ref:
- https://github.com/notch8/archives_online/issues/199